requirements/base.txt
django-filter==24.2